cloud-functions/silver_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.14.0
requests>=2.31.0

